Обновлена backend часть проекта

+++ backend/src/main/java/ru/sber/services/TaskServiceImpl.java
@@ -1,5 +1,6 @@
 package ru.sber.services;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.sber.entities.AbridgedTask;
 import ru.sber.entities.Category;
@@ -9,9 +10,12 @@ import ru.sber.repositories.TaskRepository;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
 /**
  * Сервис для взаимодействия с задачами
  */
+@Slf4j
 @Service
 public class TaskServiceImpl implements TaskService {
 
@@ -25,26 +29,24 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Long createTask(Task task) {
-        return taskRepository.save(task).getId();
-    }
-
-    @Override
-    public List<AbridgedTask> findAll(long userId) {
-        List<Category> categories = categoryService.findAllByUserId(userId);
-        if (categories.isEmpty()) {
-            return List.of();
-        } else {
-            List<AbridgedTask> taskList = new ArrayList<>();
-            for (Category category: categories) {
-                taskList.addAll(findAllByCategoryId(category.getId()));
-            }
-            return taskList;
+        boolean categoryIsRight = categoryService.checkExistence(task.getCategory().getId());
+        if (categoryIsRight) {
+            return taskRepository.save(task).getId();
         }
+        return -1L;
     }
 
     @Override
-    public List<AbridgedTask> isNotify(long userId) {
-        return categoryService.findAllByUserId(userId)
+    public List<AbridgedTask> findAll() {
+        List<Category> categories = categoryService.findAll();
+        return categories.stream()
+                .flatMap(category -> findAllByCategoryId(category.getId()).stream())
+                .toList();
+    }
+
+    @Override
+    public List<AbridgedTask> isNotify() {
+        return categoryService.findAll()
                 .stream()
                 .flatMap(category ->
                         taskRepository.findAllByCategory_Id(category.getId())
@@ -56,25 +58,35 @@ public class TaskServiceImpl implements TaskService {
 
     private static boolean isNotify(Task task) {
         LocalDateTime nowTime = LocalDateTime.now();
-        boolean isYearMatched = task.getDateAndTimeOfTask().getYear() == nowTime.getYear();
-        boolean isMonthMatched = task.getDateAndTimeOfTask().getMonth() == nowTime.getMonth();
-        boolean isDayMatched  = task.getDateAndTimeOfTask().getDayOfMonth() == nowTime.getDayOfMonth();
-        long diffTime = task.getDateAndTimeOfTask().getHour() * 60 + task.getDateAndTimeOfTask().getMinute()
+        LocalDateTime taskTime = task.getDateAndTimeOfTask();
+        if (taskTime == null) {
+            return false;
+        }
+        boolean isYearMatched = taskTime.getYear() == nowTime.getYear();
+        boolean isMonthMatched = taskTime.getMonth() == nowTime.getMonth();
+        boolean isDayMatched = taskTime.getDayOfMonth() == nowTime.getDayOfMonth();
+        long diffTime = taskTime.getHour() * 60 + taskTime.getMinute()
                 - (nowTime.getHour() * 60 + nowTime.getMinute());
         return isYearMatched && isMonthMatched && isDayMatched && diffTime < 20 && diffTime >= 0;
     }
 
     @Override
     public List<AbridgedTask> findAllByCategoryId(long categoryId) {
-        return taskRepository.findAllByCategory_Id(categoryId)
-                .stream()
+        boolean categoryIsExist = categoryService.checkExistence(categoryId);
+        List<Task> tasks = new ArrayList<>();
+        if (categoryIsExist) {
+            tasks = taskRepository.findAllByCategory_Id(categoryId);
+        }
+        return tasks.stream()
                 .map(AbridgedTask::new)
                 .toList();
     }
 
     @Override
     public boolean updateTask(Task task) {
-        if (taskRepository.existsById(task.getId())) {
+        boolean taskExist = taskRepository.existsById(task.getId());
+        boolean categoryIsRight = categoryService.checkExistence(task.getCategory().getId());
+        if (taskExist && categoryIsRight) {
             taskRepository.save(task);
             return true;
         }
@@ -84,7 +96,12 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public boolean deleteTaskById(long taskId) {
-        taskRepository.deleteById(taskId);
-        return true;
+        Optional<Task> task = taskRepository.findById(taskId);
+        if (task.isPresent() &&
+                categoryService.checkExistence(task.get().getCategory().getId())) {
+            taskRepository.deleteById(taskId);
+            return true;
+        }
+        return false;
     }
 }
